Support gcov metadata with relative paths
The original implementation worked fine when tested on a CMake project,
which passes absolute paths to the compiler. However, when gcov is used
on project that passes relative paths to the compiler (e.g. ncmpcpp
using autoconf) it needs to be run from the same directory that the
compiler ran.

Luckily the gcov notes files have a field specifying the directory used
during compilation (why does not gcov simply use this?). The gcov output
parser was reimplemented in python so that it can read the working
directory from the binary .gcno files, and so that it can specify the
working directory to use when running gcov.

Now both absolute paths (e.g. CMake projects) and relative paths (e.g.
autoconf projects) works fine.

The current implementation is a bit slower on larger projects than I
would like (TestcovRefresh takes a couple of seconds for the 30K LOC in
ncmpcpp). Part of it is because today it naively parses (and caches)
coverage information for all the metadata it finds, but another problem
is that gcov includes coverage information for every header file used,
so STL and Boost headers makes a lot of unwanted noise.

Gcov has a `--relative-only` option to ignore these system headers, but
obviously this did not anticipate e.g. CMake compiler commands using
absolute paths for everything, and I have yet to find any other options
for filtering out system headers.

Only parsing the relevant metadata on demand is a better approach, but
it is not trivial to figure out what metadata concerns what source
files. I tried extracting the filenames from the gcno files, but just
reading that data took almost as long a time as generating coverage for
everything. Function locations seemed to be early in the gcno files, so
a future optimization might be to only read those.