Run gcov with g:testcov_gcov_root as its cwd for pre GCC 9.1.0 versions
Turns out that the gcov notes metadata field for the working directory
during compilation is a fairly new feature added in [gcc:c74bd3fba35]
gcov notes files is a fairly new feature.

Fallback to run gcov with g:testcov_gcov_root as its cwd if the version
in the .gcno file indicates that it does not have the filename field. If
this cwd guess was wrong and the compilation used relative paths the
test coverage analysis will silently fail, but at least not crash.

[gcc:c74bd3fba35]: https://github.com/gcc-mirror/gcc.git
commit c74bd3fba3572d5825fcc7fabfd136b22c7981ad
Author: Martin Liska <mliska@suse.cz>
Date:   2018-05-18 14:46:32 +0200

    Print working directory to gcov files (PR gcov-profile/84846).

    2018-05-18  Martin Liska  <mliska@suse.cz>

            PR gcov-profile/84846
            * coverage.c (coverage_init): Write PWD to .gcno file.
            * doc/gcov.texi: Document how working directory is printed.
            * gcov-dump.c (dump_gcov_file): Print PWD.
            * gcov.c (output_intermediate_file): Likewise.
            (read_graph_file): Read PWD string.
            (output_lines): Print PWD.

    From-SVN: r260359