Assume gcov output/filenames are UTF-8 and ignore conversion errors
Nowadays the assumption that e.g. filenames use UTF-8 instead of e.g.
ISO-8859-1 is fairly legit. However, the gcov output is an annotated
view of the entire source code, and I occiasionally stumble across
Norwegian legacy code that includes ISO-8859-1 character.

For this plugin we are only interested in the counts, filenames and line
number from the gcov output, so rather ignore non-convertible characters
than crash on the wrong assumption that the input data is UTF-8.