Run gcov without --stdout if it does not support that option
If the gcov command does not support --stdout (e.g. GCC pre 9.1.0 or the
`llvm-cov gcov` equivalent) have it generate full path .gcov files, and
parse the content of these instead. Unfortunately gcov must run from the
working directory where the compiler were run, and it is unable to write
.gcov files any other place than in the directory it runs. Therefore
this has the side effect of littering the source tree with .gcov files.