Add support for reports from clang's SanitizerCoverage
This adds support for parsing [SanitizerCoverage] reports, and the
instrumentation for creating those can be added by doing e.g.:

  $ clang++ -g main.cpp -fsanitize=address -fsanitize-coverage=trace-pc-guard
  $ ASAN_OPTIONS=coverage=1 ./a.out # writes a.out.PID.sancov

These reports does not seem to contain the same amount of system library
instrumentation as gcov, so they are probably quite a bit faster to work
with on large projects. And having reports in JSON kind of makes them
easier to work with.

Clang also has a gcov-compatible mode for [SourceBasedCodeCoverage], but
it uses an old format that is not compatible with newer versions of the
gcov tool, and that version is harder to use because users needs a
priori knowledge about the working directory used during compilation.

[SourceBasedCodeCoverage]: https://clang.llvm.org/docs/SourceBasedCodeCoverage.html
[SanitizerCoverage]: https://clang.llvm.org/docs/SanitizerCoverage.html